all are done

+++ src/main/java/lk/ijse/dep10/app/controller/CategoryDrinkController.java
@@ -31,7 +31,7 @@ public class CategoryDrinkController extends OwnerSceneController{
     public void initialize(){
         itemName=null;
         itemCategory=null;
-        itemPrize=0;
+        itemPrize=0.0;
 
         loadData();
     }
@@ -47,7 +47,7 @@ public class CategoryDrinkController extends OwnerSceneController{
                 String name = resultSet.getString("name");
                 String category = resultSet.getString("category");
                 itemCategory=category;
-                int prize = Integer.parseInt(resultSet.getString("prize"));
+                double prize = Double.parseDouble(resultSet.getString("prize"));
                 Size size1 = lk.ijse.dep10.app.util.Size.valueOf(resultSet.getString("size"));
 
                 Item foodItem = new Item(id, category, name, size1, prize);
